requirements.txt
py-cord
docker
python-dotenv~=1.0.0
rcon